refactor: Rename `…s/package/lib/stripNullPropsFromTemplateResources.js`
