+++ lib/plugins/aws/package/index.js
@@ -2,7 +2,7 @@
 
 const path = require('path');
 const mergeCustomProviderResources = require('./lib/merge-custom-provider-resources');
-const stripNullPropsFromTemplateResources = require('./lib/stripNullPropsFromTemplateResources');
+const stripNullPropsFromTemplateResources = require('./lib/strip-null-props-from-template-resources');
 const generateArtifactDirectoryName = require('./lib/generate-artifact-directory-name');
 const generateCoreTemplate = require('./lib/generate-core-template');
 const saveServiceState = require('./lib/save-service-state');
