+++ lib/plugins/aws/package/lib/strip-null-props-from-template-resources.js
