+++ test/unit/lib/plugins/aws/package/lib/strip-null-props-from-template-resources.test.js
